NodeDefCodeCheckbox: removed Checkbox inner component
