--- webapp/components/survey/SurveyForm/nodeDefs/components/types/NodeDefCode/NodeDefCodeCheckbox/Checkbox/index.js
@@ -1,1 +1,0 @@
-export { default } from './Checkbox'
